Colocando generacion de informes

+++ ServerTime/src/com/Entidades/Registro.java
@@ -12,10 +12,10 @@ package com.Entidades;
 public class Registro {
     private String nodo; 
     private String funcion; 
-    private int tiempo_inicial;
-    private int tiempo_final;
+    private Long tiempo_inicial;
+    private Long tiempo_final;
 
-    public Registro(String nodo, String funcion, int tiempo_inicial, int tiempo_final) {
+    public Registro(String nodo, String funcion, Long tiempo_inicial, Long tiempo_final) {
         this.nodo = nodo;
         this.funcion = funcion;
         this.tiempo_inicial = tiempo_inicial;
@@ -38,19 +38,19 @@ public class Registro {
         this.funcion = funcion;
     }
 
-    public int getTiempo_inicial() {
+    public Long getTiempo_inicial() {
         return tiempo_inicial;
     }
 
-    public void setTiempo_inicial(int tiempo_inicial) {
+    public void setTiempo_inicial(Long tiempo_inicial){
         this.tiempo_inicial = tiempo_inicial;
     }
 
-    public int getTiempo_final() {
+    public Long getTiempo_final() {
         return tiempo_final;
     }
 
-    public void setTiempo_final(int tiempo_final) {
+    public void setTiempo_final(Long tiempo_final) {
         this.tiempo_final = tiempo_final;
     }
